--- ci/src/wasm_revisions/BUILD.bazel
@@ -1,7 +1,0 @@
-load("@rules_python//python:defs.bzl", "py_binary")
-
-py_binary(
-    name = "wasm_revisions",
-    srcs = ["wasm_revisions.py"],
-    deps = [    ],
-)
