--- ci/src/wasm_revisions/wasm_revisions.py
@@ -1,4 +1,0 @@
-import subprocess
-
-if __name__ == "__main__":
-    print("Hello")
